+++ cmd/docgen/docs/functions.go
@@ -6,7 +6,6 @@ import (
 	"path"
 	"strings"
 
-	"github.com/nyaruka/goflow/i18n"
 	"github.com/nyaruka/goflow/utils/jsonx"
 )
 
@@ -32,22 +31,7 @@ func (g *functionListingGenerator) Name() string {
 	return "function listing"
 }
 
-func (g *functionListingGenerator) ExtractText(items map[string][]*TaggedItem) []string {
-	msgs := make([]string, 0)
-
-	for _, funcItem := range items["function"] {
-		summary := funcItem.description[0]
-		detail := strings.TrimSpace(strings.Join(funcItem.description[1:len(funcItem.description)-1], "\n"))
-
-		msgs = append(msgs, summary)
-		if detail != "" {
-			msgs = append(msgs, detail)
-		}
-	}
-	return msgs
-}
-
-func (g *functionListingGenerator) Generate(baseDir, outputDir string, items map[string][]*TaggedItem, po *i18n.PO) error {
+func (g *functionListingGenerator) Generate(baseDir, outputDir string, items map[string][]*TaggedItem, gettext func(string) string) error {
 	funcItems := items["function"]
 	listings := make([]*functionListing, len(funcItems))
 
@@ -63,8 +47,8 @@ func (g *functionListingGenerator) Generate(baseDir, outputDir string, items map
 
 		listings[i] = &functionListing{
 			Signature: funcItem.tagValue + funcItem.tagExtra,
-			Summary:   po.GetText("", summary),
-			Detail:    po.GetText("", detail),
+			Summary:   gettext(summary),
+			Detail:    gettext(detail),
 			Examples:  examples,
 		}