+++ flows/template_test.go
@@ -24,7 +24,7 @@ func TestTemplateTranslation(t *testing.T) {
 	channel := assets.NewChannelReference("0bce5fd3-c215-45a0-bcb8-2386eb194175", "Test Channel")
 
 	for i, tc := range tcs {
-		tt := NewTemplateTranslation(types.NewTemplateTranslation(*channel, envs.Language("eng"), tc.Content, len(tc.Variables)))
+		tt := NewTemplateTranslation(types.NewTemplateTranslation(*channel, envs.Language("eng"), envs.Country("US"), tc.Content, len(tc.Variables)))
 		result := tt.Substitute(tc.Variables)
 		assert.Equal(t, tc.Expected, result, "%d: unexpected template substitution", i)
 	}
@@ -32,28 +32,66 @@ func TestTemplateTranslation(t *testing.T) {
 
 func TestTemplates(t *testing.T) {
 	channel1 := assets.NewChannelReference("0bce5fd3-c215-45a0-bcb8-2386eb194175", "Test Channel")
-	tt1 := types.NewTemplateTranslation(*channel1, envs.Language("eng"), "Hello {{1}}", 1)
-	tt2 := types.NewTemplateTranslation(*channel1, envs.Language("spa"), "Hola {{1}}", 1)
-	template := NewTemplate(types.NewTemplate("c520cbda-e118-440f-aaf6-c0485088384f", "greeting", []*types.TemplateTranslation{tt1, tt2}))
+	tt1 := types.NewTemplateTranslation(*channel1, envs.Language("eng"), envs.NilCountry, "Hello {{1}}", 1)
+	tt2 := types.NewTemplateTranslation(*channel1, envs.Language("spa"), envs.Country("EC"), "Que tal {{1}}", 1)
+	tt3 := types.NewTemplateTranslation(*channel1, envs.Language("spa"), envs.Country("ES"), "Hola {{1}}", 1)
+	template := NewTemplate(types.NewTemplate("c520cbda-e118-440f-aaf6-c0485088384f", "greeting", []*types.TemplateTranslation{tt1, tt2, tt3}))
 
 	tas := NewTemplateAssets([]assets.Template{template})
 
 	tcs := []struct {
 		UUID      assets.TemplateUUID
 		Channel   *assets.ChannelReference
-		Languages []envs.Language
+		Locales   []envs.Locale
 		Variables []string
 		Expected  string
 	}{
-		{"c520cbda-e118-440f-aaf6-c0485088384f", channel1, []envs.Language{"eng", "spa"}, []string{"Chef"}, "Hello Chef"},
-		{"c520cbda-e118-440f-aaf6-c0485088384f", channel1, []envs.Language{"deu", "spa"}, []string{"Chef"}, "Hola Chef"},
-		{"c520cbda-e118-440f-aaf6-c0485088384f", nil, []envs.Language{"deu", "spa"}, []string{"Chef"}, ""},
-		{"c520cbda-e118-440f-aaf6-c0485088384f", channel1, []envs.Language{"deu"}, []string{"Chef"}, ""},
-		{"8c5d4910-114a-4521-ba1d-bde8b024865a", channel1, []envs.Language{"eng", "spa"}, []string{"Chef"}, ""},
+		{
+			"c520cbda-e118-440f-aaf6-c0485088384f",
+			channel1,
+			[]envs.Locale{{Language: "eng", Country: "US"}, {Language: "spa", Country: "CO"}},
+			[]string{"Chef"},
+			"Hello Chef",
+		},
+		{
+			"c520cbda-e118-440f-aaf6-c0485088384f",
+			channel1,
+			[]envs.Locale{{Language: "eng", Country: ""}, {Language: "spa", Country: "CO"}},
+			[]string{"Chef"},
+			"Hello Chef",
+		},
+		{
+			"c520cbda-e118-440f-aaf6-c0485088384f",
+			channel1,
+			[]envs.Locale{{Language: "deu", Country: "DE"}, {Language: "spa", Country: "ES"}},
+			[]string{"Chef"},
+			"Hola Chef",
+		},
+		{
+			"c520cbda-e118-440f-aaf6-c0485088384f",
+			nil,
+			[]envs.Locale{{Language: "deu", Country: "DE"}, {Language: "spa", Country: "ES"}},
+			[]string{"Chef"},
+			"",
+		},
+		{
+			"c520cbda-e118-440f-aaf6-c0485088384f",
+			channel1,
+			[]envs.Locale{{Language: "deu", Country: "DE"}},
+			[]string{"Chef"},
+			"",
+		},
+		{
+			"8c5d4910-114a-4521-ba1d-bde8b024865a",
+			channel1,
+			[]envs.Locale{{Language: "eng", Country: "US"}, {Language: "spa", Country: "ES"}},
+			[]string{"Chef"},
+			"",
+		},
 	}
 
 	for _, tc := range tcs {
-		tr := tas.FindTranslation(tc.UUID, tc.Channel, tc.Languages)
+		tr := tas.FindTranslation(tc.UUID, tc.Channel, tc.Locales)
 		if tr == nil {
 			assert.Equal(t, "", tc.Expected)
 			continue